debugged: now the converter accepts floats also

+++ miles-to-km.py
@@ -2,7 +2,7 @@
 
 def function(miles):
     return 1.6*miles
-m =int(input("Enter the distance in miles:"))
+m = float(input("Enter the distance in miles:"))
 km = function(m)
 
 print(m,"miles is equal to",km,"km")
